updated overlay with some comments
